fix(10): WR-02 compute cost byte math in u64 to keep the SC1 Block guardrail correct on wasm32
`envi_compute::cost::estimate` multiplied receiver/byte counts in `usize`.
On wasm32 `usize` is 32-bit, so a large grid (e.g. ~426k receivers at
n_sub=4) wraps `tensor_bytes` to a small value in a release build (overflow
checks off); the guardrail then compares the wrapped value against
`budget_bytes` and returns Ok/Warn instead of the hard `Block` SC1 exists to
provide. Latent today because CalcPanel passes a fixed 2 GiB budget (< u32::MAX),
but it activates the moment the real OPFS quota (> 2 GiB on desktop) is wired in.

- `CostEstimate::{tensor_bytes, working_set_bytes}` are now `u64`; the products
  use `saturating_mul` so an overflowing grid saturates to `u64::MAX` and always
  trips `Block`, never wraps under budget.
- `guardrail` takes `budget_bytes: u64`; `WARN_TENSOR_BYTES` is `u64`.
- The boundary DTO already carries bytes as JS-safe `f64` (`est.tensor_bytes as f64`).
- New test `large_grid_byte_math_does_not_overflow_u32_and_still_blocks` proves a
  >u32::MAX-byte grid still Blocks against the 2 GiB budget.

+++ crates/envi-compute-wasm/src/lib.rs
@@ -177,7 +177,7 @@ pub fn estimate_cost(req: JsValue) -> Result<JsValue, JsValue> {
         req.n_sub as usize,
         req.n_workers as usize,
     );
-    let guard = envi_compute::cost::guardrail(&est, req.budget_bytes as usize);
+    let guard = envi_compute::cost::guardrail(&est, req.budget_bytes as u64);
     to_js(&CostEstimateResult {
         receiver_count: est.receiver_count as u32,
         tensor_bytes: est.tensor_bytes as f64,
